backend/devices/views.py
from rest_framework import viewsets
from devices.models import Device
from devices.serializers import DeviceSerializer

class DeviceViewSet(viewsets.ModelViewSet):
    queryset = Device.objects.all()
    serializer_class = DeviceSerializer
