backend/requirements.txt
django
djangorestframework
django-filter
napalm
drf-spectacular
requests
django-cors-headers
channels[daphne]
psycopg2-binary
channels_redis
